Fix model datatype
Bfloat16 does not work on this devcontainer, which has a relatively recent version of CUDA.

+++ src/solu_utils.py
@@ -52,7 +52,7 @@ pio.renderers.default = "notebook"
 
 
 ROOT = Path('/workspace/solu_project/')
-DTYPE = torch.bfloat16
+DTYPE = torch.float32 # torch.bfloat16 is not supported
 test_var = 1
 # %%
 for i in tqdm.tqdm(range(1)):
